in the middle of refactoring, but have no clue how to deal with lsps...
