+++ .config/nvim/.gitignore
@@ -1,0 +1,2 @@
+/.netrwhist
+/plugin
