+++ .config/nvim/lua/lsp/_sumneko_lua.lua
@@ -1,0 +1,100 @@
+-- # LUA
+--
+local util = require'lspconfig/util'
+local system_name
+if vim.fn.has("mac") == 1 then
+  system_name = "macOS"
+elseif vim.fn.has("unix") == 1 then
+  system_name = "Linux"
+elseif vim.fn.has('win32') == 1 then
+  system_name = "Windows"
+else
+  print("Unsupported system for sumneko")
+end
+-- set the path to the sumneko installation; if you previously installed via the now deprecated :LspInstall, use
+local sumneko_root_path = '/home/spicy/build/lua-language-server'
+local sumneko_binary = sumneko_root_path.."/bin/"..system_name.."/lua-language-server"
+
+return {
+  cmd = {sumneko_binary, "-E", sumneko_root_path .. "/main.lua"};
+  settings = {
+    Lua = {
+      runtime = {
+        -- Tell the language server which version of Lua you're using (most likely LuaJIT in the case of Neovim)
+        version = 'LuaJIT',
+        -- Setup your lua path
+        path = vim.split(package.path, ';'),
+      },
+      diagnostics = {
+        -- Get the language server to recognize the `vim` global
+        globals = {'vim'},
+      },
+      workspace = {
+        -- Make the server aware of Neovim runtime files
+        library = {
+          [vim.fn.expand('$VIMRUNTIME/lua')] = true,
+          [vim.fn.expand('$VIMRUNTIME/lua/vim/lsp')] = true,
+        },
+      },
+    },
+  },
+  default_config = {
+    filetypes = {'lua'};
+    root_dir = function(fname)
+      return util.find_git_ancestor(fname) or util.path.dirname(fname)
+    end;
+    log_level = vim.lsp.protocol.MessageType.Warning;
+  };
+  docs = {
+    package_json = "https://raw.githubusercontent.com/sumneko/vscode-lua/master/package.json";
+    description = [[
+    https://github.com/sumneko/lua-language-server
+    Lua language server.
+    `lua-language-server` can be installed by following the instructions [here](https://github.com/sumneko/lua-language-server/wiki/Build-and-Run-(Standalone)).
+    **By default, lua-language-server doesn't have a `cmd` set.** This is because nvim-lspconfig does not make assumptions about your path. You must add the following to your init.vim or init.lua to set `cmd` to the absolute path ($HOME and ~ are not expanded) of you unzipped and compiled lua-language-server.
+    ```lua
+    local system_name
+    if vim.fn.has("mac") == 1 then
+      system_name = "macOS"
+    elseif vim.fn.has("unix") == 1 then
+      system_name = "Linux"
+    elseif vim.fn.has('win32') == 1 then
+      system_name = "Windows"
+    else
+      print("Unsupported system for sumneko")
+    end
+    -- set the path to the sumneko installation; if you previously installed via the now deprecated :LspInstall, use
+    local sumneko_root_path = vim.fn.stdpath('cache')..'/lspconfig/sumneko_lua/lua-language-server'
+    local sumneko_binary = sumneko_root_path.."/bin/"..system_name.."/lua-language-server"
+    require'lspconfig'.sumneko_lua.setup {
+      cmd = {sumneko_binary, "-E", sumneko_root_path .. "/main.lua"};
+      settings = {
+        Lua = {
+          runtime = {
+            -- Tell the language server which version of Lua you're using (most likely LuaJIT in the case of Neovim)
+            version = 'LuaJIT',
+            -- Setup your lua path
+            path = vim.split(package.path, ';'),
+          },
+          diagnostics = {
+            -- Get the language server to recognize the `vim` global
+            globals = {'vim'},
+          },
+          workspace = {
+            -- Make the server aware of Neovim runtime files
+            library = {
+              [vim.fn.expand('$VIMRUNTIME/lua')] = true,
+              [vim.fn.expand('$VIMRUNTIME/lua/vim/lsp')] = true,
+            },
+          },
+        },
+      },
+    }
+    ```
+    ]];
+    default_config = {
+      root_dir = [[root_pattern(".git") or bufdir]];
+    };
+  };
+}
+-- vim:et ts=2
